refactor: 전역 utils 폴더에 있던 node_color 파일 node 폴더의 constants 폴더로 이동 후 color로 파일명 변경
