+++ frontend/src/features/mindmap/node/constants/colors.ts
